Fixed some warnings that the latest clang version is triggering

+++ include/cmajor/helpers/cmaj_PatchHelpers.h
@@ -488,7 +488,8 @@ inline std::optional<float> PatchParameterProperties::toValueFromDiscreteOptionI
 {
     if (auto numDiscreteSteps = getNumDiscreteOptions(); numDiscreteSteps > 1)
     {
-        auto value = static_cast<float> (minValue + ((maxValue - minValue) * i) / (numDiscreteSteps - 1));
+        auto value = static_cast<float> (minValue + ((maxValue - minValue) * static_cast<float> (i))
+                        / static_cast<float> ((numDiscreteSteps - 1u)));
 
         return shouldRound ? std::round (value) : value;
     }
